% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenProjectR.R
\name{eigenProjectR}
\alias{eigenProjectR}
\title{eigenProjectR}
\usage{
eigenProjectR(new_counts, rotation)
}
\arguments{
\item{new_counts}{raw gene count matrix (where genes are rows and samples are
columns)}

\item{rotation}{named vector where the names are genes and the values are the
prcomp function's rotation values for each genes. These are the eigenvalues.}
}
\value{
The new_counts matrix with the eigen transformation
}
\description{
This function takes in a count matrix (where genes (features) are rows and samples are
columns) as well as a named vector with the eigenvalues (see \code{\link[=run_pca]{run_pca()}})
and pulls the gene (feature) information from the rotation vector and cuts down
the new_counts matrix to match the rotation vector gene (feature) names. Any
genes (features) missing from the input new_counts matrix will be filled with
zeros.
}
\details{
The function will scale the new_counts matrix in the same manner as \code{\link[=run_pca]{run_pca()}}
and matrix multiply by the rotation vector. The matrix output is equivalent
to prcomp's "$x" matrix.
}
