% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamoRph.R
\name{extract_prcomp_scaling}
\alias{extract_prcomp_scaling}
\title{extract_prcomp_scaling}
\usage{
extract_prcomp_scaling(prcomp_object)
}
\arguments{
\item{prcomp_object}{A precomputed prcomp object}
}
\value{
a list object containing the prcomp "center" and "scale"
values for use in the metamoRph function
}
\description{
This function takes in a prcomp object and returns the center and scale
vectors a list that can be used in metamoRph::metamoRph.
}
\details{
This function is used by metamoRph::run_pca
}
