% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamoRph.R
\name{metamoRph}
\alias{metamoRph}
\title{metamoRph}
\usage{
metamoRph(
  new_counts,
  rotation,
  center_scale = NULL,
  sample_cpm_scale = TRUE,
  log1p = TRUE
)
}
\arguments{
\item{new_counts}{raw gene count matrix (where genes are rows and samples are
columns)}

\item{rotation}{matrix where the row names are genes and the col names are the
principal components. If you used metamoRph::run_pca() then this would be
in the output$PCA$rotation slot.}

\item{center_scale}{list object where the $center slot has the center values
and the $scale slot has the scale value for the "scale" function. If you do not
give a value here, then feature/gene scaling WILL NOT HAPPEN.}

\item{sample_cpm_scale}{Default is TRUE; This should match the value given
to metamoRph::run_pca(). If you are using your own rotations and don't know whether you have
done cpm scaling then...you probably should set this to FALSE.}

\item{log1p}{Default is TRUE; log1p scales the input count matrix}
}
\value{
A matrix with the transformated eigenvalue matrix which should be equivalent
to the original rotation matrix's eigenvalue/pattern matrix (The $x slot from
the output of prcomp)
}
\description{
This function takes in a count matrix (where genes (features) are rows and samples are
columns) as well as a named vector with the eigenvalues (see \code{\link[=run_pca]{run_pca()}})
and pulls the gene (feature) information from the rotation vector and cuts down
the new_counts matrix to match the rotation vector gene (feature) names. Any
genes (features) missing from the input new_counts matrix will be replaced with zeros.
}
\details{
The function will scale the new_counts matrix in the same manner as \code{\link[=run_pca]{run_pca()}}
and matrix multiply by the rotation vector. The output is equivalent
to the prcomp "$x" matrix.
}
