% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{model_apply}
\alias{model_apply}
\title{model_apply}
\usage{
model_apply(
  list_of_models,
  experiment_data,
  experiment_labels = "",
  return_predictions = FALSE
)
}
\arguments{
\item{list_of_models}{list object containing one model per sample type (e.g. photoreceptors vs not-photoreceptors)}

\item{experiment_data}{Projected data \code{\link[=metamoRph]{metamoRph()}}}

\item{experiment_labels}{Optional labels for the users experiment_data}

\item{return_predictions}{By default the predicted labels are returned.
If set to TRUE, then the entire matrix of probabilities for each sample (against
each label type) is returned.}
}
\value{
By default, a table of predicted labels for your data, if return_predictions set to TRUE then
a matrix of probabilities for each sample type is returned (where rows are samples and columns
are each sample type probability)
}
\description{
This function uses the output from \code{\link[=model_build]{model_build()}}
}
