% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{model_build}
\alias{model_build}
\title{model_build}
\usage{
model_build(
  training_data,
  training_labels,
  num_PCs = 200,
  BPPARAM = BiocParallel::SerialParam(),
  model = "lm",
  verbose = TRUE
)
}
\arguments{
\item{training_data}{sample (row) by principal component (column) matrix}

\item{training_labels}{vector which has the row-matched labels (e.g. cell types)
for each sample.}

\item{num_PCs}{number of principal components to use from the training_data.
Defaults to the first (top) 200.}

\item{BPPARAM}{The BiocParallel class}

\item{model}{lm or glm. We strongly recommend using lm.}

\item{verbose}{Print training status for each label type}
}
\value{
A list of models for each individual label type
}
\description{
This function uses the training_data and the training_labels to build a lm
based model for each label type which can be used in \code{\link[eigenProjectR:model_apply]{eigenProjectR::model_apply()}}.
The training_data is intended to be the sample x PC (principal component)
row x column matrix. Which is the \verb{$x} output from base R prcomp. We provide
precomputed prcomp PCA outputs from the \url{plae.nei.nih.gov} resource
for adult human eye, adult mouse eye, fetal human eye, and fetal mouse eye (
see \code{vignette("pca_download", package = "eigenProjectR")})
}
