% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamoRph.R
\name{normalize_data}
\alias{normalize_data}
\title{normalize_data}
\usage{
normalize_data(feature_by_sample, sample_scale = "cpm", log1p = TRUE)
}
\arguments{
\item{feature_by_sample}{features (row) by sample (column) matrix}

\item{sample_scale}{Options include \code{cpm}, \code{seurat}, \code{zscale},  and \code{none}}

\item{log1p}{Default is TRUE}
}
\value{
a list object containing the prcomp "center" and "scale"
values for use in the metamoRph function
}
\description{
Applies \emph{sample} level normalization and high variable feature (gene)
selection. If \code{cpm} scaling selected, then \code{scuttle::calculateCPM} is
applied before log1p (if \code{TRUE}).
\code{zscale} will apply \code{log1p} first (if \code{TRUE}).
}
\details{
This function is used by both metamoRph::metamoRph and metamoRph::run_pca
}
