% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenProjectR.R
\name{run_pca}
\alias{run_pca}
\title{run_pca}
\usage{
run_pca(feature_by_sample, meta, ntop = 1000, HVG_selection = "scran")
}
\arguments{
\item{feature_by_sample}{raw gene count matrix (where genes are rows and samples are
columns)}

\item{meta}{metadata for the samples. The rows must match the columns
for \code{feature_by_sample}}

\item{ntop}{Number of features to use in the prcomp PCA. Defaults to 1000.}

\item{HVG_selection}{either "classic" or "scran" to select the "ntop" features. "classic"
will simply the top n features by variance and "scran" will use the scran package's
strategy of scaling variance by expression (as highly expressed features/genes) will
also have higher variance and thus may be less useful for sample distinction}
}
\value{
named list object
}
\description{
This function takes in a count matrix (where genes (features) are rows and samples are
columns) and sample level metadata and returns a list object with a R::prcomp calculated
object, the metadata, the percent variance explained for each principal component, and
the genes (features) chosen for the PCA
}
